refactor(nextjs): extract shared preview setup to eliminate preview.tsx duplication
